Updating the prompt to use tput instead of escape codes
This commit updates the prompt to use tput instead of escape codes. This
is a more portable solution. The escape codes work on Linux, but not on
macOS. The tput commands work on both.

The tput commands are wrapped in \[ and \] to tell bash that these
commands don't take up any space on the command line.

ref: https://mug896.github.io/bash-shell/colors_and_prompt.html