Add a screenrc file to customize the screen session indicator
This commit adds a screenrc file to customize the screen session indicator.
The screenrc file is a simple file that sets the terminal capabilities
for xterm* to ti@:te@. This allows the screen session indicator to be
displayed correctly when using the screen command.

The install.sh script has been updated to append the contents of the
screenrc file to the user's .screenrc file. This will allow the screen
session indicator to be displayed correctly when using the screen command.

The prompt file has been updated to add a screen session indicator to the
prompt. The screen session indicator will display the screen session ID
if the session is running in a screen session. If the session is not
running in a screen session, the screen session indicator will not be
displayed.